--- grade-scripts/single-site-grade.js
@@ -1,110 +1,0 @@
-const puppeteer = require('puppeteer')
-const listManager = require('./shared/list-manager')
-const utils = require('./shared/utils')
-const program = require('commander')
-const chalk = require('chalk')
-
-const Grade = require('../src/classes/grade')
-const https = require('../src/https')
-const trackers = require('../src/trackers')
-const surrogates = require('../src/surrogates')
-
-let siteToCheck
-let browser
-let grade
-
-const handleRequest = (request) => {
-    let url = request.url()
-    let type = request.resourceType()
-    let upgradedUrl = https.getUpgradedUrl(url)
-
-    if (url !== upgradedUrl) {
-        url = upgradedUrl
-    }
-
-    let tracker = trackers.isTracker(url, siteToCheck, type)
-
-    if (tracker && tracker.block) {
-        grade.update({ trackerBlocked: tracker })
-
-        if (tracker.redirectUrl) {
-            url = tracker.redirectUrl
-        } else {
-            request.abort()
-            return
-        }
-    }
-
-    request.continue({
-        url: url
-    })
-}
-
-const run = async () => {
-    // load any lists and plug them into any classes that wait for them
-    await listManager.loadLists()
-    https.addLists(listManager.getList('https'))
-    trackers.addLists({
-        entityList: listManager.getList('entityList'),
-        whitelist: listManager.getList('whitelist')
-    })
-    surrogates.addLists(listManager.getList('surrogates'))
-
-    // set up headless browser
-    browser = await puppeteer.launch({
-        args: ['--no-sandbox']
-    })
-
-    const page = await browser.newPage()
-    grade = new Grade(siteToCheck)
-    let requests = []
-
-    // set useragent/viewport
-    let userAgent = await browser.userAgent()
-
-    await page.emulate({
-        // just in case some sites block headless visits
-        userAgent: userAgent.replace('Headless', ''),
-        viewport: {
-            width: 1440,
-            height: 812
-        }
-    })
-
-    await page.setRequestInterception(true)
-    page.on('request', handleRequest)
-    page.on('response', (response) => {
-        if (!utils.responseIsOK(response)) {
-            console.log(chalk.red(`got ${response.status()} for ${response.url()}`))
-        }
-    })
-
-    console.log(`getting grade for ${siteToCheck}`)
-    // wait for the page to load and then an extra 3s, to be sure
-    try {
-        await page.goto(`http://${siteToCheck}`, { timeout: 10000, waitUntil: 'load' })
-    } catch (e) {
-        console.log(chalk.red(`timed out for ${siteToCheck}`))
-    }
-    await page.waitFor(3000)
-
-    // check if https
-    if (page.url().indexOf('https:') === 0) {
-        grade.update({ hasHttps: true })
-    }
-
-    console.log(chalk.green(`grade for ${siteToCheck} is: ${JSON.stringify(grade.get())}`))
-    console.log(chalk.green(`number of trackers blocked: ${grade.totalBlocked}`))
-
-    await browser.close()
-}
-
-program
-    .arguments('<hostname>')
-    .action((hostname) => {
-        siteToCheck = hostname
-
-        run()
-    })
-    .parse(process.argv)
-
